requirements.txt
httpx==0.27.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-httpx==0.44b0
